+++ sample/app/ui-material/material-sample-form.model.ts
@@ -11,6 +11,14 @@ import {
 
 export const MATERIAL_SAMPLE_FORM_MODEL = [
 
+    new DynamicCheckboxModel({
+        id: "dynamicStuff",
+        label: "Dynamic Stuff",
+        additional: {
+            color: "primary"
+        }
+    }),
+
     new DynamicFormGroupModel({
 
         id: "stay",
@@ -70,7 +78,7 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
             }),
 
             new DynamicInputModel({
-
+                hidden: true,
                 id: "roomQuantity",
                 inputType: "number",
                 placeholder: "Room Quantity",
@@ -91,7 +99,19 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
         },
         errorMessages: {
             required: "Field is required"
-        }
+        },
+        relation: [
+            {
+                action: "DISABLE",
+                when: [
+                    {
+                        id: "dynamicStuff",
+                        operator: "===",
+                        value: true
+                    }
+                ]
+            }
+        ]
     }),
 
     new DynamicInputModel({
@@ -107,19 +127,44 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
         },
         additional: {
             color: "accent"
-        }
+        },
+        relation: [
+            {
+                action: "DISABLE",
+                when: [
+                    {
+                        id: "dynamicStuff",
+                        operator: "===",
+                        value: false
+                    }
+                ]
+            }
+        ]
     }),
 
     new DynamicInputModel({
 
         id: "email",
         placeholder: "E-Mail",
+        hidden: true,
         validators: {
             email: null
         },
         errorMessages: {
             email: "Field has no valid email"
-        }
+        },
+        relation: [
+            {
+                action: "HIDDEN",
+                when: [
+                    {
+                        id: "dynamicStuff",
+                        operator: "===",
+                        value: true
+                    }
+                ]
+            }
+        ]
     }),
 
     new DynamicInputModel({
@@ -134,7 +179,19 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
         },
         errorMessages: {
             required: "Field is required"
-        }
+        },
+        relation: [
+            {
+                action: "VISIBLE",
+                when: [
+                    {
+                        id: "dynamicStuff",
+                        operator: "===",
+                        value: true
+                    }
+                ]
+            }
+        ]
     }),
 
     new DynamicFormGroupModel({
@@ -169,7 +226,7 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
                 id: "state",
                 hint: "Autocomplete",
                 placeholder: "State",
-                list: ['Alabama', 'Alaska', 'American Samoa', 'Arizona', 'Arkansas', 'California', 'Colorado', 'Connecticut', 'Delaware', 'District of Columbia', 'Federated States of Micronesia', 'Florida', 'Georgia', 'Guam', 'Hawaii', 'Idaho', 'Illinois', 'Indiana', 'Iowa', 'Kansas', 'Kentucky', 'Louisiana', 'Maine', 'Marshall Islands', 'Maryland', 'Massachusetts', 'Michigan', 'Minnesota', 'Mississippi', 'Missouri', 'Montana', 'Nebraska', 'Nevada', 'New Hampshire', 'New Jersey', 'New Mexico', 'New York', 'North Carolina', 'North Dakota', 'Northern Mariana Islands', 'Ohio', 'Oklahoma', 'Oregon', 'Palau', 'Pennsylvania', 'Puerto Rico', 'Rhode Island', 'South Carolina', 'South Dakota', 'Tennessee', 'Texas', 'Utah', 'Vermont', 'Virgin Island', 'Virginia', 'Washington', 'West Virginia', 'Wisconsin', 'Wyoming']
+                list: ["Alabama", "Alaska", "American Samoa", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", "Delaware", "District of Columbia", "Federated States of Micronesia", "Florida", "Georgia", "Guam", "Hawaii", "Idaho", "Illinois", "Indiana", "Iowa", "Kansas", "Kentucky", "Louisiana", "Maine", "Marshall Islands", "Maryland", "Massachusetts", "Michigan", "Minnesota", "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada", "New Hampshire", "New Jersey", "New Mexico", "New York", "North Carolina", "North Dakota", "Northern Mariana Islands", "Ohio", "Oklahoma", "Oregon", "Palau", "Pennsylvania", "Puerto Rico", "Rhode Island", "South Carolina", "South Dakota", "Tennessee", "Texas", "Utah", "Vermont", "Virgin Island", "Virginia", "Washington", "West Virginia", "Wisconsin", "Wyoming"]
             }),
 
             new DynamicInputModel({